// workspaces/litellm-orchestrator/plugins/litellm-orchestrator/src/schema/openapi/generated/apis/Api.client.ts
export type TypedResponse<T> = Omit<Response, 'json'> & {
  json: () => Promise<T>;
};

/**
 * Options you can pass into a request for additional information.
 *
 * @public
 */
export interface RequestOptions {
  token?: string;
}
/**
 * @public
 */
export type CreateUser = {
  body: CreateUserRequest;
};
/**
 * @public
 */
export type CreateUserKey = {
  path: {
    userId: string;
  };
  body: any;
};
/**
 * @public
 */
export type DeleteUserKey = {
  path: {
    userId: string;
    keyId: string;
  };
};
/**
 * @public
 */
export type GetUser = {
  path: {
    userId: string;
  };
};
/**
 * @public
 */
export type GetUserKeys = {
  path: {
    userId: string;
  };
};

/**
 * @public
 */
export class DefaultApiClient {
  private readonly discoveryApi: DiscoveryApi;
  private readonly fetchApi: FetchApi;

  constructor(options: {
    discoveryApi: { getBaseUrl(pluginId: string): Promise<string> };
    fetchApi?: { fetch: typeof fetch };
  }) {
    this.discoveryApi = options.discoveryApi;
    this.fetchApi = options.fetchApi || { fetch: crossFetch };
  }

  /**
   * @param createUserRequest -
   */
  public async createUser(
    // @ts-ignore
    request: CreateUser,
    options?: RequestOptions,
  ): Promise<TypedResponse<User>> {
    const baseUrl = await this.discoveryApi.getBaseUrl(pluginId);

    const uriTemplate = `/users`;

    const uri = parser.parse(uriTemplate).expand({});

    return await this.fetchApi.fetch(`${baseUrl}${uri}`, {
      headers: {
        'Content-Type': 'application/json',
        ...(options?.token && { Authorization: `Bearer ${options?.token}` }),
      },
      method: 'POST',
      body: JSON.stringify(request.body),
    });
  }

  /**
   * @param userId -
   * @param body -
   */
  public async createUserKey(
    // @ts-ignore
    request: CreateUserKey,
    options?: RequestOptions,
  ): Promise<TypedResponse<UserKey>> {
    const baseUrl = await this.discoveryApi.getBaseUrl(pluginId);

    const uriTemplate = `/users/{userId}/keys`;

    const uri = parser.parse(uriTemplate).expand({
      userId: request.path.userId,
    });

    return await this.fetchApi.fetch(`${baseUrl}${uri}`, {
      headers: {
        'Content-Type': 'application/json',
        ...(options?.token && { Authorization: `Bearer ${options?.token}` }),
      },
      method: 'POST',
      body: JSON.stringify(request.body),
    });
  }

  /**
   * @param userId -
   * @param keyId -
   */
  public async deleteUserKey(
    // @ts-ignore
    request: DeleteUserKey,
    options?: RequestOptions,
  ): Promise<TypedResponse<void>> {
    const baseUrl = await this.discoveryApi.getBaseUrl(pluginId);

    const uriTemplate = `/users/{userId}/keys/{keyId}`;

    const uri = parser.parse(uriTemplate).expand({
      userId: request.path.userId,
      keyId: request.path.keyId,
    });

    return await this.fetchApi.fetch(`${baseUrl}${uri}`, {
      headers: {
        'Content-Type': 'application/json',
        ...(options?.token && { Authorization: `Bearer ${options?.token}` }),
      },
      method: 'DELETE',
    });
  }

  /**
   * @param userId -
   */
  public async getUser(
    // @ts-ignore
    request: GetUser,
    options?: RequestOptions,
  ): Promise<TypedResponse<User>> {
    const baseUrl = await this.discoveryApi.getBaseUrl(pluginId);

    const uriTemplate = `/users/{userId}`;

    const uri = parser.parse(uriTemplate).expand({
      userId: request.path.userId,
    });

    return await this.fetchApi.fetch(`${baseUrl}${uri}`, {
      headers: {
        'Content-Type': 'application/json',
        ...(options?.token && { Authorization: `Bearer ${options?.token}` }),
      },
      method: 'GET',
    });
  }

  /**
   * @param userId -
   */
  public async getUserKeys(
    // @ts-ignore
    request: GetUserKeys,
    options?: RequestOptions,
  ): Promise<TypedResponse<UserKeyList>> {
    const baseUrl = await this.discoveryApi.getBaseUrl(pluginId);

    const uriTemplate = `/users/{userId}/keys`;

    const uri = parser.parse(uriTemplate).expand({
      userId: request.path.userId,
    });

    return await this.fetchApi.fetch(`${baseUrl}${uri}`, {
      headers: {
        'Content-Type': 'application/json',
        ...(options?.token && { Authorization: `Bearer ${options?.token}` }),
      },
      method: 'GET',
    });
  }
}