// workspaces/litellm-orchestrator/plugins/litellm-orchestrator/src/schema/openapi/generated/models/User.model.ts
//

// ******************************************************************
// * THIS IS AN AUTOGENERATED FILE. DO NOT EDIT THIS FILE DIRECTLY. *
// ******************************************************************

/**
 * A user in the system
 * @public
 */
export interface User {
  userId: string;
  createdAt: string;
  maxBudget: number;
  budgetDuration: string;
  spend: number;
}